moved modal to button and refactored onKeyPress as it is deprecated
