+++ src/components/vizToolbar/components/button/button.js
@@ -1,11 +1,21 @@
-import React from "react"
+import React, { useState } from "react"
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { 
   faUndoAlt, faRedoAlt, faHistory, faSyncAlt, faShareAlt, faInfoCircle, 
 } from '@fortawesome/free-solid-svg-icons'
 import * as set from "../../toolBarConfig/toolBarConfig"
+import Modal from "../../../modal/modal"
 
 export default function Button(props) {
+  // controls display of embedded content details
+  const [detailModal, setDetailModal] = useState(false);
+
+  // toggles display of modal
+  const handleModal = () => {
+    console.log('detailModal', detailModal)
+    setDetailModal(!detailModal);
+  }
+
   const btnStyles = `${props.color} ${props.outline} ${props.rounded}`;
   
   // array of supported buttons, all of them are displayed by default
@@ -15,7 +25,7 @@ export default function Button(props) {
     {name: "Reset", icon: faHistory, function: () => {props.vizObj.revertAllAsync()}},
     {name: "Refresh", icon: faSyncAlt, function: () => {props.vizObj.refreshDataAsync()}},
     {name: "Details", icon: faInfoCircle, function: () => {
-      props.handleModal();
+      handleModal();
       console.log('getName', props.vizObj.getWorkbook().getName())
       console.log('getActiveSheet', props.vizObj.getWorkbook().getActiveSheet())
       console.log('getPublishedSheetsInfo', props.vizObj.getWorkbook().getPublishedSheetsInfo())
@@ -44,6 +54,26 @@ export default function Button(props) {
           </button>
         ))
       )}
+      <Modal
+        card={true}
+        display={detailModal}
+        setDisplay={handleModal}
+        title="Sample Test Modal"
+        footer={
+          <>
+            <button className="button is-primary">Save changes</button>
+            <button className="button">Cancel</button>
+          </>
+        }
+      >
+        <p>
+          Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut 
+          labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris 
+          nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit 
+          esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in 
+          culpa qui officia deserunt mollit anim id est laborum.
+        </p>
+      </Modal>
     </>
   )
 }
